+++ nhspy/nhspy.py
@@ -1,1 +1,16 @@
 # -*- coding: utf-8 -*-
+"""
+# nhspy : Implementation of nhspy
+
+Summary :
+    Implements a common core of functionality for use on NHS projects
+Use Case :
+    As a Developer I want a standard common core of functionality So that I can develop applications which behave in a
+    consistent manner.
+
+Testable Statements :
+    Can I Input, Manipulate and Output date/times in a form that is
+                                    compliant with the CUI Date format (dd-mm-YYYY HH:MM) - See cui.py
+    ....
+"""
+