+++ HISTORY.rst
@@ -6,3 +6,11 @@ History
 ------------------
 
 * First release on PyPI.
+
+
+0.2.0 (2016-05-22)
+------------------
+
+* Implementation of the `cui.DateTime`_ class.
+
+.. _`cui.DateTime`: CUI/Date_Time